requirements.txt
requests
python-dotenv
pytrends
